relay/requirements.txt
Flask>=3.0
PyNaCl>=1.5
